Rename _internal.code_quality to _internal.typing
